Add utility function to hide/show Cursor
This adds a genric utility module to provide for utility function, which
should not be part of the main code.

The first utility function hides the cursor during "tqdm" operations, in
which the cursor might interfere with the display format. It also makes
sure, the cursor is shown again before returning to the command line to
avoid loss of cursor on the command line.